git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@14240 01821946-851b-0410-b7f4-f1477c70e57a

+++ gameio/createmesh.cpp
@@ -1718,7 +1718,7 @@ if (nEdge >= 0) {
 #endif
 		int32_t nEdges = gameData.segData.nEdges;
 		do {
-			if (nEdges >= edges.Length () - 1) {
+			if (nEdges >= (int32_t) edges.Length () - 1) {
 				if (!edges.Resize (nEdges * 2))
 					return -1;
 				}
@@ -1739,7 +1739,7 @@ if (nEdge >= 0) {
 	}
 else {
 	i = 0;
-	if ((nEdges >= edges.Length ()) && !edges.Resize (nEdges * 2))
+	if ((nEdges >= (int32_t) edges.Length ()) && !edges.Resize (nEdges * 2))
 		return -1;
 	CMeshEdge *pEdge = &edges [nEdges++];
 	pEdge->m_nVertices [0] = nVertex1;
